Refactor policies agent API into modules
