+++ examples/multi_agent_human_chat/agents/policies/agent/routes/__init__.py
@@ -1,0 +1,7 @@
+"""Routes for the Policies Agent API."""
+
+from libraries.vespa import VespaClient
+
+vespa_client = VespaClient()
+
+__all__ = ["vespa_client"]
